uploading php file to get tables from db
This file will be accessed using jquery, allowing for server-side interaction after the page has loaded.

At the moment, it just takes the Problem table and converts it into a html table string. Hopefully it shouldn't be too much work to make it do any table and return table strings.