cleaned up and changed functions
I've cleaned up some of the placeholder text and function that was in the base draft, and changed the getTable function into a spearate file. This allows for me to use and access it after the page has loaded. Instead I have a jQuery function that will return a string with the html of a table. This allows easy refreshing.